feat: restaura layout funcional e implementa seletor de data

+++ convex/tarefas.ts
@@ -1,43 +1,54 @@
 import { v } from "convex/values";
 import { mutation, query } from "./_generated/server";
 
-// busca todas as tarefas
 export const listar = query({
-    args: {},
-    handler: async (ctx) => {
-        return await ctx.db.query("tarefas").collect();
-    },
+  args: {},
+  handler: async (ctx) => {
+    return await ctx.db
+      .query("tarefas")
+      .order("desc")
+      .collect();
+  },
 });
 
-// adiciona uma nova tarefa
 export const adicionar = mutation({
-    args: {
-        assunto: v.string(),
-        prioridade: v.string()
-    },
-    handler: async (ctx, args) => {
-        await ctx.db.insert("tarefas", {
-            assunto: args.assunto,
-            prioridade: args.prioridade,
-            finalizada: false, // todas as tarefas começam como não finalizadas.
-        });
-    },
+  args: {
+    assunto: v.string(),
+    prioridade: v.string(),
+    prazo: v.string(),
+  },
+  handler: async (ctx, args) => {
+    const prazoFinal = args.prazo.trim() === "" ? "Sem prazo" : args.prazo;
+    await ctx.db.insert("tarefas", {
+      assunto: args.assunto,
+      prioridade: args.prioridade,
+      finalizada: false,
+      prazo: prazoFinal,
+    });
+  },
 });
 
-// marca como concluída ou desmarcar 
 export const alternarStatus = mutation({
-    args: { id: v.id("tarefas"), atual: v.boolean() },
-    handler: async (ctx, args) => {
-      await ctx.db.patch(args.id, {
-        finalizada: !args.atual, // Inverte o valor booleano
-      });
-    },
-  });
+  args: { 
+    id: v.id("tarefas"), 
+    atual: v.boolean() 
+  },
+  handler: async (ctx, args) => {
+    await ctx.db.patch(args.id, { 
+      finalizada: !args.atual 
+    });
+  },
+});
 
-  // remove uma tarefa do banco
 export const remover = mutation({
-    args: { id: v.id("tarefas") }, // usamos o ID para saber qual apagar
-    handler: async (ctx, args) => {
+  args: { 
+    id: v.id("tarefas") 
+  },
+  handler: async (ctx, args) => {
+    const tarefaExistente = await ctx.db.get(args.id);
+    
+    if (tarefaExistente) {
       await ctx.db.delete(args.id);
-    },
-  });
+    }
+  },
+});
